pyproject.toml
[project]
name = "dbt-airflow-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
   "apache-airflow>=2.10.3",
   "apache-airflow-providers-postgres>=6.2.3",
   "dbt-core>=1.8.9",
   "dbt-postgres>=1.8.2",
   "astronomer-cosmos>=1.7.1",
   "fastapi>=0.104.0",
   "uvicorn[standard]>=0.24.0",
   "asyncpg>=0.29.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.6.0",
    "ruff>=0.8.2",
    "mypy>=1.13.0",
]

emacs = [
    "debugpy>=1.8.9",
    "jedi-language-server>=0.42.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "B",   # flake8-bugbear
    "I",   # isort
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "UP",  # pyupgrade
    "RUF"  # ruff-specific rules
]

[tool.ruff.lint.isort]
known-first-party = ["app"]
known-third-party = ["typer", "rich"]
combine-as-imports = true
split-on-trailing-comma = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.12"
strict = true
implicit_reexport = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = ["tests/*"]
follow_imports = "silent"
show_column_numbers = true

[[tool.mypy.overrides]]
module = "airflow.*"
ignore_missing_imports = true
